!998 feat(#I37WAM): PetaPoco extensions support PageQuery * feat: 增加分页查询扩展 * refactor: QueryPageOptions 增加 IsPage 参数用于区分是否分页 * refactor: QueryData 参数 TotalCount 更改为 long 类型

+++ src/BootstrapBlazor/Components/ListView/ListViewBase.cs
@@ -78,7 +78,7 @@ namespace BootstrapBlazor.Components
         /// <summary>
         /// 获得/设置 数据总条目
         /// </summary>
-        protected int TotalCount { get; set; }
+        protected long TotalCount { get; set; }
 
         /// <summary>
         /// 获得/设置 当前页码
