+++ src/BootstrapBlazor/Components/Table/Table.razor.Pagination.cs
@@ -50,7 +50,7 @@ namespace BootstrapBlazor.Components
         /// <summary>
         /// 获得/设置 数据总条目
         /// </summary>
-        protected int TotalCount { get; set; }
+        protected long TotalCount { get; set; }
 
         /// <summary>
         /// 获得/设置 当前页码
